Move mealplan components into mealplan directory
